Move some SPI types to the SPI package
This commit moves some SPI types to the SPI package. Breaking apart the internals in a way that avoids circular dependencies between the failsafe and spi packages is challenging. The solution is to use more interfaces instead of structs, and have types in either package depend on their own interface, which in a few cases just happen to implement the same methods.